+++ src/orquestra/sdk/_base/cli/_dorq/_ui/_presenters.py
@@ -20,6 +20,7 @@ from tabulate import tabulate
 from orquestra.sdk._base import _config, _dates, _env, _services, serde
 from orquestra.sdk._base._dates import Instant
 from orquestra.sdk.schema import responses
+from orquestra.sdk.schema.configs import ConfigName, RuntimeConfiguration, RuntimeName
 from orquestra.sdk.schema.ir import ArtifactFormat
 from orquestra.sdk.schema.workflow_run import (
     TaskInvocationId,
@@ -210,6 +211,7 @@ class LoginPresenter:
         )
 
     def prompt_config_saved(self, url, config_name):
+        # TODO: tell the user what runtime is being used.
         click.echo("Token saved in config file.")
         click.echo(f"Configuration name for {url} is {config_name}")
 
@@ -220,6 +222,56 @@ class LoginPresenter:
 
     def open_url_in_browser(self, url) -> bool:
         return webbrowser.open(url)
+
+
+class ConfigPresenter:
+    """
+    Present config information to the user.
+    """
+
+    def print_configs_list(
+        self,
+        configs: t.Sequence[RuntimeConfiguration],
+        status: t.Mapping[ConfigName, bool],
+        message: t.Optional[str] = "Stored configs:",
+    ):
+        """
+        Print a list of stored configs.
+        """
+        click.echo(message)
+        click.echo(
+            tabulate(
+                [
+                    [
+                        # show config name
+                        click.style(config.config_name, bold=True),
+                        #
+                        # show runtime name, colour coded blue for CE and green for QE
+                        click.style(config.runtime_name, fg="blue")
+                        if config.runtime_name == RuntimeName.CE_REMOTE
+                        else click.style(config.runtime_name, fg="green"),
+                        #
+                        # show cluster URI
+                        config.runtime_options["uri"],
+                        #
+                        # show a green tick if the token is current, and a red cross if
+                        # it is not.
+                        click.style("\u2713", fg="green")
+                        if status[config.config_name]
+                        else click.style("\u2A09", fg="red"),
+                    ]
+                    for config in configs
+                ],
+                colalign=("left",),
+                tablefmt="plain",
+                headers=[
+                    click.style("Config Name", underline=True),
+                    click.style("Runtime", underline=True),
+                    click.style("Server URI", underline=True),
+                    click.style("Current Token", underline=True),
+                ],
+            ),
+        )
 
 
 def _format_datetime(dt: t.Optional[Instant]) -> str: